Finished the controller test cases [Martin&Aaron]

+++ src/main/java/com/tw/apistackbase/service/CompanyService.java
@@ -3,12 +3,9 @@ package com.tw.apistackbase.service;
 import com.tw.apistackbase.core.Company;
 import com.tw.apistackbase.repository.CompanyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-
+import static java.util.Objects.isNull;
 import static org.springframework.data.domain.PageRequest.of;
 
 @Service
@@ -16,6 +13,10 @@ public class CompanyService {
 
     @Autowired
     private CompanyRepository companyRepository;
+
+    public Iterable<Company> findAll() {
+        return companyRepository.findAll();
+    }
 
     public Iterable<Company> findAllWithPageRequest(int page, int size) {
         return companyRepository.findAll(of(page, size));
@@ -25,29 +26,28 @@ public class CompanyService {
         return companyRepository.findByNameContaining(name);
     }
 
-    public HttpEntity save(Company company) {
-        companyRepository.save(company);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    public Company save(Company company) {
+        return companyRepository.save(company);
     }
 
-    public HttpEntity update(Company company, String name) {
-        company.setName(name);
-        companyRepository.save(company);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public Company update(String name, Long id) {
+        Company company = findById(id);
+        if (!isNull(company)) {
+            company.setName(name);
+            companyRepository.save(company);
+        }
+        return company;
     }
 
-    public Company findById(Long id) {
+    public Company deleteById(Long id) {
+        Company company = findById(id);
+        if (!isNull(company)) {
+            companyRepository.deleteById(id);
+        }
+        return company;
+    }
+
+    private Company findById(Long id) {
         return companyRepository.findById(id).orElse(null);
     }
-
-    public HttpEntity deleteById(Long id) {
-        companyRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
 }
-    
-
-
-
-
-
